requirements.txt
# List required packages in this file, one per line.
numpy
matplotlib
pandas
scipy
lmfit
spec2nexus
ipython
numpydoc
sphinx_copybutton
suitcase-csv
suitcase-json-metadata
databroker>=1.2.2
databroker-pack
lmfit
xraylarch
pyresttable
area-detector-handlers